+++ frontend/src/hooks/Main/useCouponDetail.ts
@@ -1,5 +1,5 @@
 import { useQuery } from 'react-query';
-import { requestInstance } from '../../apis/axios';
+import { client } from '../../apis/axios';
 import { API_PATH } from '../../constants/api';
 
 const formatingRawDate = date => date.split('T')[0] + ' ' + date.split('T')[1].split('.')[0];
@@ -8,7 +8,7 @@ export const useReservationDetail = couponId => {
   const { data, isLoading, isError } = useQuery(
     ['reservationDetail', couponId],
     async () => {
-      const { data } = await requestInstance({
+      const { data } = await client({
         method: 'get',
         url: `${API_PATH.GET_COUPON_DETAIL(couponId)}`,
       });
@@ -30,7 +30,7 @@ export const useReservationDetail = couponId => {
 
 export const useNotUsedCouponDetail = couponId => {
   const { data, isLoading, isError } = useQuery(['notUsedDetail', couponId], async () => {
-    const { data } = await requestInstance({
+    const { data } = await client({
       method: 'get',
       url: `${API_PATH.GET_COUPON_DETAIL(couponId)}`,
     });
