+++ frontend/src/hooks/Main/useMain.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useQuery } from 'react-query';
-import { requestInstance } from '../../apis/axios';
+import { client } from '../../apis/axios';
 import { API_PATH } from '../../constants/api';
 import { Coupon, CouponType } from '../../types';
 
@@ -14,7 +14,7 @@ const useMain = () => {
   const [currentType, setCurrentType] = useState<CouponType>('entire');
 
   const { data, isLoading, error } = useQuery<Coupon[]>('coupon', async () => {
-    const { data } = await requestInstance({
+    const { data } = await client({
       method: 'get',
       url: `${API_PATH.RECEIVED_COUPONS_NOT_USED}`,
     });
